requirements.txt
pandas
requests
langchain
pydantic
openai
langchain-openai 
python-dotenv
langgraph
langsmith
langchain_anthropic
termcolor
docker
langchain_community
chromadb
bs4
gcloud
setuptools
ansi2html
httplib2
google-api-python-client
pymongo
langchain-google-vertexai
praw
mem0ai